chore: Refactored network_config_data into a module and clustered related classes
